// app/main/enums/settings-keys.ts
enum SettingsKeys {
  HostIpAddress = 'HOST_IP_ADDRESS',
  ContentMode = 'CONTENT_MODE ',
  MessageboardStatus = 'MESSAGEBOARD_STATUS',
  MessageToneEnabled = 'MESSAGE_TONE_ENABLED',
  PostDeliverySchedule = 'POST_DELIVERY_SCHEDULE',
  SystemRebootTime = 'SYSTEM_REBOOT_TIME',
  SyncTimeStatus = 'SYNC_TIME_STATUS',
  SyncDisplayOffStatus = 'SYNC_DISPLAYOFF_STATUS',

  //
  DeviceSerialNo = 'DEVICE_SERIAL_NO',
  IsEnableOption = 'IS_ENABLE_OPTION',
  HostPcLastConnection = 'HOST_PC_LAST_CONNECTION',
  CurrentDeliveryName = 'CURRENT_DELIVERY_NAME',
  CurrentTimetableName = 'CURRENT_TIMETABLE_NAME',
  CurrentContentName = 'CURRENT_CONTENT_NAME',
  CurrentMode = 'CURRENT_MODE',
  AlertContentName = 'ALERT_CONTENT_NAME',

  IsDisplayOff = 'IS_DISPLAY_OFF',

  InsertionStartDatetime = 'INSERTION_START_DATETIME',
  InsertionEndDatetime = 'INSERTION_END_DATETIME',

}

export default SettingsKeys;